+++ src/components/gameOfLife.css
